updates to add missing comment_promote_log table, update numerical capacity for Firehose srcid column
tracks internal commit 7cc019206d4853a646e08bf8b5d794b09d5fd0ad

+++ plugins/FireHose/mysql_schema.sql
@@ -19,7 +19,7 @@ CREATE TABLE firehose (
 	is_spam enum('no', 'yes') default 'no',
 	primaryskid smallint(5) default '0',
 	tid smallint(6) default '0',
-	srcid mediumint(8) unsigned NOT NULL default '0',
+	srcid int(8) unsigned NOT NULL default '0',
 	url_id mediumint(8) unsigned NOT NULL default '0',
 	toptags varchar(255) default '',
 	email varchar(255) NOT NULL default '',
